Cat/Prod v2.1

+++ src/Producte.java
@@ -2,6 +2,7 @@ import java.util.ArrayList;
 
 
 /**
+ * @Class Producte
  * Representa un producte del sistema, amb nom, llista de similituts 
  * i index que representa la seva posicio al Catalge (@see Cataleg)
  * @author Alejandro Lorenzo Navarro
@@ -9,9 +10,13 @@ import java.util.ArrayList;
  */
 public class Producte {
 
+    /** Index que concideix amb index on es troba el producte dins de cataleg*/
     public int index;
-    public String nom; 
-    public ArrayList<Int> llista_Similituds;
+    /** String que representa el nom unic que se li ha donat al producte */
+    public String nom;
+    /** ArrayList que representa les similituds de aquest producte amb la resta
+     * cada similitud amb un producte es troba a index corresponent al index*/
+    public ArrayList<Double> llista_Similituds;
 
     /**
      * @param inde El index a Cataleg_Productes de la classe Cataleg on es troba el producte 
@@ -19,7 +24,7 @@ public class Producte {
      * @param llista Una llista amb totes les similituds del productes
      * Descripció: Es crea una nova instancia de productes
      */
-    public Producte(int index, String nom, ArrayList<Int> llista) {
+    public Producte(int index, String nom, ArrayList<Double> llista) {
         this.index = index;
         this.nom = nom;
         this.llista_Similituds = llista;
